initial snakemake workflow refactoring
